Added support for the root folder containing a dot
The auto-generation of strongly typed localization files was filtering out .resx files that had more than a single dot in the filename but this was done using the path, so all localization files where being excluded because of the dot on the folder name. This makes it apply that filter only on the actual filename.